test: move test Hardware table to conftest.py
